pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.1",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
youtube-mcp = "main:main"

[tool.hatch.build.targets.wheel]
sources = ["src"]
only-include = ["src"]

[tool.hatch.build.targets.sdist]
include = [
  "src/**",
  "README.md",
  "pyproject.toml"
]
